Resampling 2x Speedup
Many bigger and smaller improvements, altogether making sinc about twice as fast as before.
- sinc resampling is now multithreaded
- sinc core cleaned up and documented better
- end padding is now already trimmed in preparation phase, so both lin and sinc resamplers end correctly
- signals are now used if they have already been loaded
- evaluated using lowpass coefficient array - result is inferior due to interpolation, currently old system remains in use
- evaluated parallel processing with guvectorize: slower
- tried cuda: not working, would need much more effort

+++ util/io_ops.py
@@ -4,7 +4,7 @@ import os
 def read_file(file_path):
 	print("Reading",file_path)
 	soundob = sf.SoundFile(file_path)
-	return soundob.read(always_2d=True), soundob.samplerate, soundob.channels
+	return soundob.read(always_2d=True, dtype="float32"), soundob.samplerate, soundob.channels
 
 def write_file(file_path, signal, sr, channels, suffix="_out"):
 	# write the final signal
